Development

+++ database/migrations/2018_08_14_191111_create_contents_table.php
@@ -14,7 +14,7 @@ class CreateContentsTable extends Migration
     public function up()
     {
         Schema::create('contents', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('type');
             $table->string('image')->nullable();
@@ -22,13 +22,13 @@ class CreateContentsTable extends Migration
             $table->text('url')->nullable();
             $table->text('yt_url')->nullable();
 
-            $table->unsignedInteger('folder_id');
+            $table->unsignedBigInteger('folder_id');
             $table->foreign('folder_id')
                 ->references('id')
                 ->on('folders')
                 ->onDelete('cascade');
 
-            $table->unsignedInteger('content_id')->nullable();
+            $table->unsignedBigInteger('content_id')->nullable();
             $table->foreign('content_id')
                 ->references('id')
                 ->on('contents')
