+++ database/migrations/2019_01_02_185709_create_tag_poi_pivot_table.php
@@ -16,7 +16,7 @@ class CreateTagPoiPivotTable extends Migration
         Schema::create('tag_poi', function (Blueprint $table) {
             $table->tag();
 
-            $table->unsignedInteger('poi_id');
+            $table->unsignedBigInteger('poi_id');
             $table->foreign('poi_id')
                 ->references('id')
                 ->on('pois')
