+++ database/migrations/2019_01_02_195158_create_tag_folder_pivot_table.php
@@ -16,7 +16,7 @@ class CreateTagFolderPivotTable extends Migration
         Schema::create('tag_folder', function (Blueprint $table) {
             $table->tag();
 
-            $table->unsignedInteger('folder_id');
+            $table->unsignedBigInteger('folder_id');
             $table->foreign('folder_id')
                 ->references('id')
                 ->on('folders')
