+++ database/migrations/2019_01_02_201931_create_tag_content_pivot_table.php
@@ -16,7 +16,7 @@ class CreateTagContentPivotTable extends Migration
         Schema::create('tag_content', function (Blueprint $table) {
             $table->tag();
 
-            $table->unsignedInteger('content_id');
+            $table->unsignedBigInteger('content_id');
             $table->foreign('content_id')
                 ->references('id')
                 ->on('contents')
