+++ database/migrations/2019_01_29_092727_create_tag_template_pivot_table.php
@@ -16,7 +16,7 @@ class CreateTagTemplatePivotTable extends Migration
         Schema::create('tag_template', function (Blueprint $table) {
             $table->tag();
 
-            $table->unsignedInteger('template_id');
+            $table->unsignedBigInteger('template_id');
             $table->foreign('template_id')
                 ->references('id')
                 ->on('templates')
